qdrant_db/requirements.txt
qdrant-client==1.12.1
openai==1.55.3
langchain==0.3.9
python-dotenv==1.0.1
